% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.DateYMD}
\alias{as.DateYMD}
\title{Format Y-M-D as date}
\usage{
as.DateYMD(x)
}
\arguments{
\item{x}{character string containing the date in the format "\%Y-\%m-\%d"}
}
\description{
Format Y-M-D as date
}
