% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.mismatch.R.R
\name{check.mismatch.R}
\alias{check.mismatch.R}
\title{Check mismatch between \R versions}
\usage{
check.mismatch.R(date)
}
\arguments{
\item{date}{character string  (yyyy-mm-dd), or date value, with the date which determines the
version of the package, and all dependencies, to be loaded (and installed if needed).}
}
\description{
Warn if there is a mismatch between the current and required \R version
}
\seealso{
\code{\link[=r.version.check]{r.version.check()}} to get a named list containing the version
numbers of the \R version currently in use and required.
}
