% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.toc.R
\name{cross.toc}
\alias{cross.toc}
\title{Show toc table with multiple packages}
\usage{
cross.toc(pkgs, date1 = "1970-1-1", date2 = Sys.Date())
}
\arguments{
\item{pkgs}{character vector containing the package names.}

\item{date1, date2}{date range to consider (in the format "\%Y-\%m-\%d").}
}
\value{
A \code{data.frame} with 3 columns:
\describe{
\item{Version}{The package version number}
\item{Published}{The date at which the specific version was published}
\item{Package}{The package name}
}
}
\description{
Show toc table with multiple packages
}
\examples{
cross.toc(c("magrittr", "rlang"), date1 = "2012-02-01", date2 = "2020-02-01")

}
\seealso{
\code{\link[=toc]{toc()}} for the same function for a single package.
}
