% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disable.local.R
\name{disable.local}
\alias{disable.local}
\title{Disable all packages in your local (non-groundhog) library to avoid conflicts when loading via groundhog.library()}
\usage{
disable.local()
}
\description{
To avoid version conflicts between the version of the package \code{groundhog} is trying
to load, and the version R Studio loads automatically from the local library, you can
disable all packages in the default local library. This action is fully and instantaneously
reversible, but if you rely on groundhog to load packages you may not need to ever reverse it.
To re-enable all packages simply  run \code{enable.local()}
}
\details{
Packages are stored into folders inside the main folder consisting of the local folder.
You can find your default local library running \code{.libPaths()[1]}.
When you run \code{disable.local()} every folder inside the library is renamed, adding the string
\verb{_DISABLED}. This prevents R Studio from loading those packages automatically. The packages are still
listed when running \code{installed.packages()} but they are not actually available to be used (e.g.,
using \code{library()} will fail to load them. When the command \code{enable.local()} is run, all folders are renamed back, eliminated the term \verb{_DISABLED}
and they are again loadable by R Studio and by users with the \code{library()} command
}
\seealso{
\code{\link[=enable.local]{enable.local()}}
}
