% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disable.packages.R
\name{enable.packages}
\alias{enable.packages}
\title{Re-enable previously disabled packages in your non-groundhog library}
\usage{
enable.packages(packages)
}
\description{
This function reverses the actions taken by \code{disable.packages()}
}
\seealso{
\code{\link[=disable.packages]{disable.packages()}}
}
