% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.seconds.left.R
\name{estimate.seconds.left}
\alias{estimate.seconds.left}
\title{Estimate seconds left in install}
\usage{
estimate.seconds.left(k, start.time, snowball)
}
\arguments{
\item{k}{numeric. Current step in the snowball process.}

\item{start.time}{the result of \code{\link[=Sys.time]{Sys.time()}} when the installation process
started}

\item{snowball}{the result of \code{\link[=get.snowball]{get.snowball()}}.}
}
\value{
the remaining time left in seconds
}
\description{
Estimate seconds left in install
}
