% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{explain.day}
\alias{explain.day}
\title{Show explanation for how and why groundhog suggests a date to use for loading all packages in a script}
\usage{
explain.day()
}
\description{
Show explanation for how and why groundhog suggests a date to use for loading all packages in a script
}
