% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.active.R
\name{get.active}
\alias{get.active}
\title{Get Active packages as data.frame()}
\usage{
get.active()
}
\value{
$pkg has the pkg name, $pkg_vrs the pkg_vrs
}
\description{
Get Active packages as data.frame()
}
