% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.dependencies.R
\name{get.dependencies}
\alias{get.dependencies}
\title{Get dependencies for one package on a given date}
\usage{
get.dependencies(pkg, date, include.suggests = FALSE)
}
\description{
Get direct dependencies for one package \code{pkg}.
For CRAN packages it uses the date when it was published on CRAN
For GitHub and Gitlab package it is based on the last commit
pushed on, or prior to the the \code{date}.
}
