% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.groundhog.folder.R
\name{get.groundhog.folder}
\alias{get.groundhog.folder}
\title{Get current local path to groundhog folder}
\usage{
get.groundhog.folder()
}
\value{
the path to the groundhog folder, the meta-library where
\code{\link[=groundhog.library]{groundhog.library()}} downloads and stores packages that can be loaded
}
\description{
Get current local path to groundhog folder
}
\note{
you can change the location of this folder with the command
\code{set.groundhog.folder("path")}.
}
\examples{
\dontrun{
get.groundhog.folder()
}

}
\seealso{
\code{\link[=set.groundhog.folder]{set.groundhog.folder()}}
}
