% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.groundhogr.folder.R
\name{get.groundhogr.folder}
\alias{get.groundhogr.folder}
\title{Get groundhogR folder location}
\usage{
get.groundhogr.folder()
}
\value{
the path to the groundhogR folder where groundhogR files will be
stored and where packages loaded with \code{\link[=groundhog.library]{groundhog.library()}} will be
installed.
}
\description{
Get groundhogR folder location
}
\note{
you can change the location of this folder by editing the environment
variable GROUNDHOG_FOLDER. In R, you can do this with the command
\code{set.groundhogr.folder("path")}.
}
\examples{
get.groundhogr.folder()

}
\seealso{
\code{\link[=set.groundhogr.folder]{set.groundhogr.folder()}}
}
