% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundhog.unload.R
\name{get.parents}
\alias{get.parents}
\title{Load \code{cran.toc}}
\usage{
get.parents(dep12, dep0)
}
\arguments{
\item{pkg}{character, name of package to be unoaded}
}
\value{
(invisibly) \code{TRUE}/\code{FALSE} depending on the success/failure of this
operation.
}
\description{
Unload a package upon request, unloading all its parents (reverse-dependencies)
}
