% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pkg_search_paths.R
\name{get.pkg_search_paths}
\alias{get.pkg_search_paths}
\title{Determine R Paths to search for an already installed package (versions of same minor, in reverse chronological order)}
\usage{
get.pkg_search_paths(pkg, vrs)
}
\description{
Determine R Paths to search for an already installed package (versions of same minor, in reverse chronological order)
}
