% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R.pkg.date.R
\name{get.R.pkg.date}
\alias{get.R.pkg.date}
\title{Get date of binary package from MRAN}
\usage{
get.R.pkg.date(pkg_vrs, R_vrs)
}
\arguments{
\item{pkg_vrs}{character. The required package and its version separated by
an underscore.}

\item{R_vrs}{character. The required R version.}
}
\value{
the date of the binary the download from MRAN, in the format
"\%Y-\%m-\%d".
}
\description{
Get date of binary package from MRAN, given R version being used
}
\examples{
\donttest{
}

}
\seealso{
\code{\link[=get.version]{get.version()}} for the opposite operation: determining the package
version from a given date.
}
