% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.snowball.R
\name{install.snowball}
\alias{install.snowball}
\title{8 Function 8 - Install snowball}
\usage{
install.snowball(
  pkg,
  date,
  include.suggests,
  force.install = FALSE,
  force.source = FALSE,
  plot.console = TRUE,
  quiet.install = TRUE,
  current.deps = "Rcpp"
)
}
\arguments{
\item{include.suggests}{Logical. Should suggested packages be installed?}

\item{force.install}{Logical (defaults to \code{FALSE}). If \code{TRUE}, even if
package is found for this R-build, it will reinstall it.}

\item{force.source}{Logical (defaults to \code{FALSE}). If \code{TRUE},` will skip CRAN
and MRAN attempts, download tar.gz, and install from it.}

\item{plot.console}{Logical (defaults to \code{TRUE}). Should information on
installation time left be printed in plots dialogue when installing from
source.}

\item{quiet.install}{Logical (defaults to \code{TRUE}).Run \code{\link[=install.packages]{install.packages()}}
with \code{quiet = TRUE}?}

\item{current.deps}{Dependencies that will install current version regardless
of date.}
}
\description{
8 Function 8 - Install snowball
}
