% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.source.R
\name{install.source}
\alias{install.source}
\title{Install package from source}
\usage{
install.source(pkg_vrs, lib, date, force.download = FALSE, quiet = FALSE)
}
\arguments{
\item{lib}{
    character vector giving the library directories where to
    install the packages.  Recycled as needed.  If missing, defaults to
    the first element of \code{\link{.libPaths}()}.
  }

\item{quiet}{
    logical: if true, reduce the amount of output.
  }
}
\description{
Install package from source
}
