% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.feedback.R
\name{installation.feedback}
\alias{installation.feedback}
\title{Installation feedback}
\usage{
installation.feedback(k, date, snowball, start.time)
}
\arguments{
\item{k}{numeric. Current step in the snowball process.}

\item{date}{date (yyyy-mm--dd) for version that determines version of package to be loaded}

\item{snowball}{the result of \code{\link[=get.snowball]{get.snowball()}}.}

\item{start.time}{the result of \code{\link[=Sys.time]{Sys.time()}} when the installation process
started}
}
\description{
Print progress about the package installation process (number of packages
already installed or remaining, time left, etc.)
}
