% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.pkg_vrs.installed}
\alias{is.pkg_vrs.installed}
\title{Is pkg_vrs installed (within same R-minor version)?}
\usage{
is.pkg_vrs.installed(pkg, vrs)
}
\arguments{
\item{pkg}{character. Name of the package to install}

\item{vrs}{character. Version of the package to install.}
}
\description{
Is pkg_vrs installed (within same R-minor version)?
}
