% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.cran.toc.R
\name{load.cran.toc}
\alias{load.cran.toc}
\title{Load the dataframe listing all CRAN packages, their dependencies and Publication date}
\usage{
load.cran.toc(update.toc = FALSE)
}
\description{
Load the dataframe listing all CRAN packages, their dependencies and Publication date
}
