% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.groundhog.R
\name{meta.groundhog}
\alias{meta.groundhog}
\title{Load a specific version of groundhog, as available on a given date}
\usage{
meta.groundhog(date)
}
\arguments{
\item{date}{character string  (yyyy-mm-dd), or date value, with the date which determines the
version of groundhog to load}
}
\description{
Load a specific version of groundhog, as available on a given date
}
\examples{
\dontrun{
#Load groundhog as available on 2021-03-12 (v1.3.2)
meta.groundhog("2021-03-12")
}

}
