% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mran.is.up.R
\name{mran.is.up}
\alias{mran.is.up}
\title{Re-enable groundhog.library() to try MRAN binaries}
\usage{
mran.is.up()
}
\description{
When an install from MRAN fails, groundhog does not try MRAN again for 5 hours
assuming the server is down. You can over-rule this preventive measure by running
this function.
}
