% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R.switch.R
\name{msg.R.switch}
\alias{msg.R.switch}
\title{Message about \R version switch}
\usage{
msg.R.switch(date)
}
\arguments{
\item{date}{date (yyyy-mm--dd) for version that determines version of package to be loaded}
}
\value{
(invisibly) the full \R version in use at \code{date}.
}
\description{
This message is printed when the \R version you currently use does not match
the \R version in use at \code{date}.
}
\seealso{
\code{\link[=r.version.check]{r.version.check()}}
}
