% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researchbox_functions.R
\name{researchbox_code}
\alias{researchbox_code}
\alias{researchbox_data}
\alias{researchbox_other}
\title{Get researchbox project paths}
\usage{
researchbox_code()

researchbox_data()

researchbox_other()
}
\value{
\itemize{
\item For \code{researchbox_code()}, the path to the \code{Code} folder of the researchbox
\item For \code{researchbox_data()}, the path to the \code{Data} folder of the researchbox
\item For \code{researchbox_other()}, the path to the \code{Other} folder of the
researchbox
}
}
\description{
Get paths relative to the different subfolders of a
\href{https://researchbox.org/}{researchbox}.
}
\examples{
x <- researchbox_other()
\dontrun{
setwd(researchbox_data())
}

}
