% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researchbox_functions.R
\name{researchbox_data}
\alias{researchbox_data}
\title{Get path to data directory in researchbox structure}
\usage{
researchbox_data()
}
\value{
Path to the \code{Data} folder within  \href{https://researchbox.org/}{researchbox} structure
}
\description{
\href{https://researchbox.org/}{researchbox} organizes files into a main folder with 5 subfolders:
code, data, other, pre-registration, and materials. This function allows referencing the
data folder if such structure is used, whereby the current r script is assumed to be in the code
folder, and the data in a neighboring folder.
For example, if locally the script is saved into
\verb{c:/dropbox/researchbox_1821/code/script.r}, then researchbox_data() will return
\verb{c:/dropbox/researchbox_1821/data/} allowing making a dynamic reference to the data path.
That way, when a user downloads the ResearchBox files to their folder, no matter where the that is locally,
because subdirectories will maintain their relative position, code that uses researchbox_data() to set the
data path will run in the new computer without needed adjustments.
If the user who downloaded the the files saves them into, say, "d:/temp/researchbox/code/script.r"
the command \code{researchbox_data()} will output "d:/temp/researchbox/data/" allowing the same code to load the
same data despite the different parental folder structure.
}
\examples{
\dontrun{
#assign dynamically path as a sibling to current script's path
data_path <- researchbox_data()           
#name of data file
file_name <- 'example.csv'                
#full path, dynamically generated
file_path <- paste0(data_path, file_name) 
data1 <- read.csv(file_path)
}

}
