% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researchbox_functions.R
\name{researchbox_dir}
\alias{researchbox_dir}
\title{Get path to parent directory in  \href{https://researchbox.org/}{ResearchBox} structure
See \code{researchbox_data()}}
\usage{
researchbox_dir()
}
\description{
Get path to parent directory in  \href{https://researchbox.org/}{ResearchBox} structure
See \code{researchbox_data()}
}
