% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/researchbox_functions.R
\name{researchbox_other}
\alias{researchbox_other}
\title{Get path to 'other' directory in  \href{https://researchbox.org/}{ResearchBox} structure
See \code{researchbox_data()}}
\usage{
researchbox_other()
}
\description{
Get path to 'other' directory in  \href{https://researchbox.org/}{ResearchBox} structure
See \code{researchbox_data()}
}
