% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.library.R
\name{restore.library}
\alias{restore.library}
\title{Restore version of packages in personal library replaced by groundhog.}
\usage{
restore.library(days)
}
\description{
When groundhog installs a package, it saves two copies of it.
One goes on the stable groundhog library (to find its location: \code{get.groundhog.folder()})
the other goes to the default personal library (to find its location: \code{.libPaths()[1]}.
Because the personal library can only hold one version of a package, groundhog replaces the
version in the personal library, but makes a copy of it in groundhog's stable library.
To restore to your default personal library the version
of packages it had prior to using groundhog, run \code{restore.library()} and the backup
copy of the original package version will be copied to the personal library.
}
\examples{
\dontrun{
restore.library()
}

}
