% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.groundhog.folder.R
\name{set.groundhog.folder}
\alias{set.groundhog.folder}
\title{Set groundhog folder location}
\usage{
set.groundhog.folder(path)
}
\arguments{
\item{path}{Character. The path to the groundhog folder containing the library
where packages are downloaded and installed.}
}
\value{
(invisibly) \code{TRUE} upon success.
}
\description{
Set groundhog folder location
}
\examples{
\dontrun{
set.groundhog.folder("~/.R_groundhog")
}

}
\seealso{
\code{\link[=get.groundhog.folder]{get.groundhog.folder()}}
}
