% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundhog.options.R
\name{set.groundhog.option}
\alias{set.groundhog.option}
\title{Set an option for groundhog}
\usage{
set.groundhog.option(name, value)
}
\arguments{
\item{name}{Character. The name of the option being set.}

\item{value}{Character or logical. The value to assign the option to.}
}
\description{
Set an option for groundhog
}
\details{
Current options that can be set manually include the following:

\link{os}: operating system. It is supposed to be detected automatically, and then
used for finding binaries for that OS. The accepted options are
'windows','mac', 'mac_arm' (for the x1+ chip). If you are using a different
operating system it is coded as 'unknown' and always installs from source.
If you are using a windows or mac machine, you can set it by hand with this function.

\link{download.sequentially}. To speed up the download of binaries, they are
by default downloaded with 'libcurl' simultaneously, if 'libcurl' is detected.
A user can manually turn this off to avoid possible errors. This will lead to
downloading packages one a time, slowing down installations by a few seconds per
package (& dependency).
}
\examples{
\dontrun{
set.groundhog.option('download.sequentially',TRUE)
}

}
\seealso{
\code{\link[=get.groundhog.option]{get.groundhog.option()}}
}
