% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.groundhogr.folder.R
\name{set.groundhogr.folder}
\alias{set.groundhogr.folder}
\title{Set groundhogR folder location}
\usage{
set.groundhogr.folder(path)
}
\arguments{
\item{path}{Character. The path to the groundhogR folder where groundhogR
files will be stored and where packages loaded with \code{\link[=groundhog.library]{groundhog.library()}}
will be installed.}
}
\description{
Set groundhogR folder location
}
\note{
This setting can also be achieved by manually editing the \code{.Renviron}
file. You can set this globally by editing \verb{~/.Renviron} or only for a
specific project by editing the \code{.Renviron} file at the root of your
project.
}
\examples{
\dontrun{
set.groundhogr.folder()
}

}
\seealso{
\code{\link[=get.groundhogr.folder]{get.groundhogr.folder()}}
}
