% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc.R
\name{toc}
\alias{toc}
\title{Show table of contents (toc) (package versions and publication dates) for a pkg, sorted chronologically}
\usage{
toc(pkg, dependencies = FALSE)
}
\description{
Show table of contents (toc) (package versions and publication dates) for a pkg, sorted chronologically
}
