% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cran.toc_if.needed.R
\name{update_cran.toc_if.needed}
\alias{update_cran.toc_if.needed}
\title{update cran.toc() if needed}
\usage{
update_cran.toc_if.needed(date)
}
\arguments{
\item{date}{In the format "\%Y-\%m-\%d" (ISO 8601 convention)}
}
\description{
update cran.toc() if needed
}
