% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.snowball.conflict.R
\name{check.snowball.conflict}
\alias{check.snowball.conflict}
\title{Check snowball conflict}
\usage{
check.snowball.conflict(snowball)
}
\arguments{
\item{snowball}{the result of \code{\link[=get.snowball]{get.snowball()}}.}
}
\description{
Check snowball conflict
}
