% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disable.local.R
\name{enable.local}
\alias{enable.local}
\title{Avoid package version conflicts by temporarily disabling packages in the local (non-groundhog) library}
\usage{
enable.local()
}
\description{
This function reverses the actions taken by \code{disable.local()}
}
\seealso{
\code{\link[=disable.local]{disable.local()}}
}
