% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.seconds.left.R
\name{estimate.seconds.left}
\alias{estimate.seconds.left}
\title{Estimate seconds left in install}
\usage{
estimate.seconds.left(k, start.time, snowball)
}
\description{
Estimate seconds left in install
}
