% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.active.R
\name{get.active}
\alias{get.active}
\title{Get active packages}
\usage{
get.active()
}
\value{
a \code{data.frame} with two columns:
\describe{
\item{pkg}{the package name}
\item{pkg_vrs}{the package name and its version, separated by an
underscore}
}
}
\description{
Get currently active (=loaded) packages in the session, with their version
}
\examples{
\donttest{
get.active()
}

}
