% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.date.for.install.binary.R
\name{get.date.for.install.binary}
\alias{get.date.for.install.binary}
\title{Get date of binary package from MRAN}
\usage{
get.date.for.install.binary(pkg_vrs)
}
\arguments{
\item{pkg_vrs}{character. The required package and its version separated by
an underscore.}
}
\value{
the date of the binary the download from MRAN, in the format
"\%Y-\%m-\%d".
}
\description{
Get date of binary package from MRAN
}
\examples{
get.date.for.install.binary("magrittr_1.0.1")

}
\seealso{
\code{\link[=get.version]{get.version()}} for the opposite operation: determining the package
version from a given date.
}
