% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.dependencies.R
\name{get.dependencies}
\alias{get.dependencies}
\title{Get dependencies for ONE package}
\usage{
get.dependencies(pkg, date, include.suggests = FALSE)
}
\arguments{
\item{include.suggests}{Logical. Should suggested packages be installed?}
}
\description{
Get dependencies for ONE package
}
