% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.groundhog.folder.R
\name{get.groundhog.folder}
\alias{get.groundhog.folder}
\title{Get groundhog folder location}
\usage{
get.groundhog.folder()
}
\value{
the path to the groundhog folder, the meta-library where
\code{\link[=groundhog.library]{groundhog.library()}} downloads and stores binaries and source files
}
\description{
Get groundhog folder location
}
\note{
you can change the location of this folder by editing the environment
variable GROUNDHOG_FOLDER. In R, you can do this with the command
\code{set.groundhog.folder("path")}.
}
\examples{
\dontrun{
get.groundhog.folder()
}

}
\seealso{
\code{\link[=set.groundhog.folder]{set.groundhog.folder()}}
}
