% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.groundhogr.folder.R
\name{get.groundhog.folder}
\alias{get.groundhog.folder}
\title{Get groundhog folder location}
\usage{
get.groundhog.folder()
}
\value{
the path to the groundhog folder where groundhogR files will be
stored and where packages loaded with \code{\link[=groundhog.library]{groundhog.library()}} will be
installed.
}
\description{
Get groundhog folder location
}
\note{
you can change the location of this folder by editing the environment
variable GROUNDHOG_FOLDER. In R, you can do this with the command
\code{set.groundhog.folder("path")}.
}
\examples{
get.groundhog.folder()
}
\seealso{
\code{\link[=set.groundhog.folder]{set.groundhog.folder()}}
}
