% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundhog.options.R
\name{get.groundhog.option}
\alias{get.groundhog.option}
\title{Get groundhog option}
\usage{
get.groundhog.option(name)
}
\description{
Get groundhog option
}
\examples{
\dontrun{
get.groundhog.option("download.sequentially")
}

}
\seealso{
\code{\link[=set.groundhog.option]{set.groundhog.option()}}
}
