% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.installed_path.R
\name{get.installed_path}
\alias{get.installed_path}
\title{Check if package version is installed for current or past R versions,}
\usage{
get.installed_path(pkg, vrs)
}
\description{
Check if package version is installed for current or past R versions,
}
