% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.installed_path.R
\name{get.installed_path}
\alias{get.installed_path}
\title{Check if package version is installed for current or past R versions,}
\usage{
get.installed_path(pkg, vrs)
}
\arguments{
\item{pkg}{character. Name of the package to install}

\item{vrs}{character. Version of the package to install.}
}
\value{
If the package with the specific version number is installed, the
installation path is returned. Otherwise, an empty string (\code{""}) is
returned.
}
\description{
Check if package version is installed for current or past R versions,
}
\examples{
\donttest{
groundhog:::get.installed_path("magrittr", "1.5")
}

}
