% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.snowball.R
\name{get.snowball}
\alias{get.snowball}
\title{Generates dataframe with all dependencies needed to install a package, with version and source of installation}
\usage{
get.snowball(pkg, date, include.suggests = FALSE, force.source = FALSE)
}
\arguments{
\item{pkg}{character string pr vector with name of target package(s) to load (and install if needed),
need not be in quotes.}

\item{date}{character string  (yyyy-mm-dd), or date value, with the date which determines the
version of the package, and all dependencies, to be loaded (and installed if needed).}

\item{include.suggests}{logical, defaults to \code{FALSE}. When set to \code{TRUE}, installs/
loads dependencies classified in the DESCRIPTION file as \code{suggested}.}

\item{force.source}{logical (defaults to \code{FALSE}). When set to \code{TRUE}, will not attempt
installing binary from CRAN or MRAN and instead download source file and install it.}
}
\description{
Generates dataframe with all dependencies needed to install a package, with version and source of installation
}
