% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.snowball.R
\name{get.snowball}
\alias{get.snowball}
\title{Generates dataframe with all dependencies needed to install a package, with version and source of installation}
\usage{
get.snowball(
  pkg,
  date,
  include.suggests = FALSE,
  force.source = FALSE,
  current.deps = .pkgenv$current.deps
)
}
\arguments{
\item{pkg}{character string with name of target package to install,
need not be in quotes.}

\item{date}{character string, or date value, with the date which determines the
version of the package, and all dependencies, to be installed. It's in yyyy-mm-dd format.}

\item{include.suggests}{logical, defaults to FALSE, when set to TRUE installs/
loads dependencies classified in the DESCRIPTION file as suggested.}

\item{force.source}{Logical (defaults to \code{FALSE}). If \code{TRUE},` will not attempt
installing binary from CRAN or MRAN and instead download source file and install it.}

\item{current.deps}{character vector with dependencies that, if needed, will be
installed with the newest version between the requested date the release of the
version of R being used, instead of necessarily with the version implied by the
date. Default includes series of C++ related packages which tend to be incompatible
across R versions.}
}
\description{
Generates dataframe with all dependencies needed to install a package, with version and source of installation
}
