% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.snowball.R
\name{get.snowball}
\alias{get.snowball}
\title{Figure out in which order dependencies should be installed to avoid conflicts}
\usage{
get.snowball(
  pkg,
  date,
  include.suggests = FALSE,
  force.source = FALSE,
  current.deps = "Rcpp"
)
}
\arguments{
\item{pkg}{character. Name of the package to install}

\item{date}{date (in the format "\%Y-\%m-\%d"), Required date for the package.}

\item{include.suggests}{Logical (defaults to \code{FALSE}. Should suggested
packages be installed?}

\item{force.source}{Logical (defaults to \code{FALSE}). If \code{TRUE},` will skip CRAN
and MRAN attempts, download tar.gz, and install from it.}

\item{current.deps}{Dependencies that will install current version regardless
of date.}
}
\description{
Figure out in which order dependencies should be installed to avoid conflicts
}
\examples{
get.snowball("magrittr", "2018-02-12")

}
