% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.snowball.R
\name{get.snowball}
\alias{get.snowball}
\title{Generates dataframe with all dependencies needed to install a package, with version and source of installation}
\usage{
get.snowball(
  pkg,
  date,
  include.suggests = FALSE,
  force.source = FALSE,
  current.deps = get.current.deps()
)
}
\arguments{
\item{pkg}{character with the name of the package to be loaded}

\item{date}{date (yyyy-mm--dd) for version that determines version of package to be loaded}

\item{include.suggests}{logical (defaults to \code{FALSE}). If \code{TRUE} will also load packages that are
suggested in the DESCRIPTION file}

\item{force.source}{Logical (defaults to \code{FALSE}). If \code{TRUE},` will skip CRAN
and MRAN attempts, download tar.gz, and install from it.}

\item{current.deps}{Character or character vector listing packages that will
be loaded/installed based on the date when the version of \R being used was
released, rather than the value of \code{date}.}
}
\description{
Generates dataframe with all dependencies needed to install a package, with version and source of installation
}
