% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.version.R
\name{get.version}
\alias{get.version}
\title{Get version from date}
\usage{
get.version(
  pkg,
  date,
  current.deps = c("Rcpp", "RcppArmadillo", "BH", "RcppEigen", "StanHeaders",
    "RcppParallel", "RcppProgress")
)
}
\arguments{
\item{pkg}{character. Name of the package to install}

\item{date}{date (in the format "\%Y-\%m-\%d"), Required date for the package.}

\item{current.deps}{Character or character vector listing all packages for
which the installed version will be the latest version available when your
\R version came out, regardless of the value for \code{date}.}
}
\description{
Get version from date
}
\examples{
\donttest{
groundhogR:::get.version("magrittr", "2018-02-12")
}

}
\seealso{
\code{\link[=get.R.pkg.date]{get.R.pkg.date()}} for the opposite operation: determining date
for a given version.
}
