% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.version.R
\name{get.version}
\alias{get.version}
\title{Get version from date}
\usage{
get.version(pkg, date, current.deps = c("Rcpp"))
}
\arguments{
\item{current.deps}{Dependencies that will install current version regardless
of date.}
}
\description{
Get version from date
}
