% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.version.R
\name{get.version}
\alias{get.version}
\title{Get version from date}
\usage{
get.version(
  pkg,
  date,
  current.deps = .pkgenv$current.deps,
  patch = c("current", "max")
)
}
\arguments{
\item{pkg}{character string with name of target package to install,
need not be in quotes.}

\item{date}{character string, or date value, with the date which determines the
version of the package, and all dependencies, to be installed. It's in yyyy-mm-dd format.}

\item{current.deps}{character vector with dependencies that, if needed, will be
installed with the newest version between the requested date the release of the
version of R being used, instead of necessarily with the version implied by the
date. Default includes series of C++ related packages which tend to be incompatible
across R versions.}

\item{patch}{Either \code{"current"} for the exact current version of the \code{pkg}
at the specified date, or \code{"max"} for the version with the same major and
minor but maximal patch version numbers.}
}
\description{
Get version from date
}
\seealso{
\code{\link[=get.R.pkg.date]{get.R.pkg.date()}} for the opposite operation: determining date
for a given version.
}
