% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundhog.library.R
\name{groundhog.library}
\alias{groundhog.library}
\title{Install packages as available on set date -  groundhog.library()}
\usage{
groundhog.library(
  pkg,
  date,
  quiet.install = TRUE,
  include.suggests = FALSE,
  current.deps = default.current.deps(),
  ignore.deps = c(),
  force.source = FALSE,
  force.install = FALSE
)
}
\arguments{
\item{pkg}{character string with name of target package to install,
need not be in quotes.}

\item{date}{character string, or date value, with the date which determines the
version of the package, and all dependencies, to be installed. It's in yyyy-mm-dd format.}

\item{quiet.install}{a logical, defaults to TRUE. When set to FALSE displays
output generated by install.packages() when installing source packages.}

\item{include.suggests}{logical, defaults to FALSE, when set to TRUE installs/
loads dependencies classified in the DESCRIPTION file as suggested.}

\item{current.deps}{character vector with dependencies that, if needed, will be
installed with the newest version between the requested date the release of the
version of R being used, instead of necessarily with the version implied by the
date. Default includes series of C++ related packages which tend to be incompatible
across R versions (to see the set of packages run 'groundhog:::default.current.deps()')}

\item{ignore.deps}{an optional character vector containing dependencies which
may be already loaded in the R session and even if the loaded version does not match
the version implied by the entered date, groundhog.library() will proceed and ignore
this conflict. If one version of a package is loaded, and a different is needed for
groundhog, the default behavior is to stop the request and ask the user to restart
the R session to unload all packages. This will bypass that behavior.}

\item{force.source}{Logical (defaults to \code{FALSE}). If \code{TRUE},` will not attempt
installing binary from CRAN or MRAN and instead download source file and install it.}

\item{force.install}{Logical (defaults to \code{FALSE}). If \code{TRUE},` will deleted
existing package files in groundhog folder, and install anew.}
}
\value{
a character vector containing all active packages for the session,
with their version number, under the format \code{pkg_vrs}.
}
\description{
Install packages as available on set date -  groundhog.library()
}
\examples{
\dontrun{
groundhog.library("magrittr", "2018-02-12")
}

}
