% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundhog.library.R
\name{groundhog.library}
\alias{groundhog.library}
\title{Install packages as available on set date -  groundhog.library()}
\usage{
groundhog.library(
  pkg,
  date,
  quiet.install = TRUE,
  include.suggests = FALSE,
  current.deps = .pkgenv$current.deps,
  ignore.deps = c(),
  force.source = FALSE,
  force.install = FALSE
)
}
\arguments{
\item{pkg}{character string with name of target package to install}

\item{date}{date (yyyy-mm--dd) for version that determines version of package to be loaded}

\item{ignore.deps}{optional character vector containing dependencies which
may mismatch those implied by the entered date and be tolerated. This will
prevent the installation to stop and request restarting the R session for
specified dependencies.

@param quiet.install  logical, defaults to TRUE,  when set to FALSE displays
output generated by install.packages() when installing source packages

@param include.suggests logical, defaults to FALSE, when set to TRUE installs/
loads dependencies classified in the DESCRIPTION file as suggested

@param current.deps character vector with dependencies that, if needed, will be
installed with the newest version between the requested date the release of the
version of R being used}

\item{force.source}{Logical (defaults to \code{FALSE}). If \code{TRUE},` will skip CRAN
and MRAN attempts, download tar.gz, and install from it.}

\item{force.install}{Logical (defaults to \code{FALSE}). If \code{TRUE}, even if
package is found for this R-build, it will reinstall it.}
}
\value{
a character vector containing all active packages for the session,
with their version number, under the format \code{pkg_vrs}.
}
\description{
Install packages as available on set date -  groundhog.library()
}
\examples{
\dontrun{
groundhog.library("magrittr", "2018-02-12")
}

}
