% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundhog.library.R
\name{groundhog.library}
\alias{groundhog.library}
\title{8  Final function:   groundhog.library()}
\usage{
groundhog.library(
  pkg,
  date,
  quiet.install = TRUE,
  plot.console = TRUE,
  include.suggests = FALSE,
  current.deps = "Rcpp",
  ignore.package.conflicts = FALSE,
  force.source = FALSE,
  force.install = FALSE
)
}
\arguments{
\item{pkg}{character. Name of the package to install}

\item{date}{date (in the format "\%Y-\%m-\%d"), Required date for the package.}

\item{quiet.install}{Logical (defaults to \code{TRUE}).Run \code{\link[=install.packages]{install.packages()}}
with \code{quiet = TRUE}?}

\item{plot.console}{logical (defaults to \code{TRUE}). Should detailed feedback
about the installation process be printed in the plot console.}

\item{include.suggests}{Logical (defaults to \code{FALSE}. Should suggested
packages be installed?}

\item{current.deps}{Dependencies that will install current version regardless
of date.}

\item{ignore.package.conflicts}{Logical (defaults to \code{FALSE}). With \code{TRUE},
if an already attached package will be installed, it is detached, but all
depednecies are left installed. With default, \code{FALSE}, script stops and
asks for session restart.}

\item{force.source}{Logical (defaults to \code{FALSE}). If \code{TRUE},` will skip CRAN
and MRAN attempts, download tar.gz, and install from it.}

\item{force.install}{Logical (defaults to \code{FALSE}). If \code{TRUE}, even if
package is found for this R-build, it will reinstall it.}
}
\value{
a character vector containing all active packages for the session,
with their version number, under the format \code{pkg_vrs}.
}
\description{
8  Final function:   groundhog.library()
}
\examples{
groundhog.library("magrittr", "2018-02-12")

}
