% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groundhog.library.R
\name{groundhog.library}
\alias{groundhog.library}
\title{Install packages as available on set date -  groundhog.library()}
\usage{
groundhog.library(
  pkg,
  date,
  quiet.install = TRUE,
  include.suggests = FALSE,
  current.deps = .pkgenv$current.deps,
  ignore.deps = c(),
  ignore.package.conflicts = FALSE,
  force.source = FALSE,
  force.install = FALSE
)
}
\arguments{
\item{date}{date (in the format "\%Y-\%m-\%d"), Required date for the package.}

\item{quiet.install}{Logical (defaults to \code{TRUE}).Run \code{\link[=install.packages]{install.packages()}}
with \code{quiet = TRUE}?}

\item{ignore.deps}{optional character vector containing dependencies which
may mismatch those implied by the entered date and be tolerated. This will
prevent the installation to stop and request restarting the R session for
specified dependencies.}

\item{force.source}{Logical (defaults to \code{FALSE}). If \code{TRUE},` will skip CRAN
and MRAN attempts, download tar.gz, and install from it.}

\item{force.install}{Logical (defaults to \code{FALSE}). If \code{TRUE}, even if
package is found for this R-build, it will reinstall it.}
}
\value{
a character vector containing all active packages for the session,
with their version number, under the format \code{pkg_vrs}.
}
\description{
Install packages as available on set date -  groundhog.library()
}
\examples{
\dontrun{
groundhog.library("magrittr", "2018-02-12")
}

}
