% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.snowball.R
\name{install.snowball}
\alias{install.snowball}
\title{Install snowball}
\usage{
install.snowball(
  snowball,
  date,
  force.install = FALSE,
  force.source = FALSE,
  quiet.install = TRUE
)
}
\arguments{
\item{snowball}{the result of \code{\link[=get.snowball]{get.snowball()}}.}

\item{date}{date (yyyy-mm--dd) for version that determines version of package to be loaded}

\item{force.install}{Logical (defaults to \code{FALSE}). If \code{TRUE}, even if
package is found for this R-build, it will reinstall it.}

\item{force.source}{Logical (defaults to \code{FALSE}). If \code{TRUE},` will skip CRAN
and MRAN attempts, download tar.gz, and install from it.}
}
\value{
Invisible \code{NULL}.
}
\description{
Install given \code{snowball} from CRAN binary, MRAN binary, CRAN source or MRAN source as needed
}
\seealso{
\code{\link[=get.snowball]{get.snowball()}} to determine in which order packages are installed
}
