% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.snowball.R
\name{install.snowball}
\alias{install.snowball}
\title{Install snowball}
\usage{
install.snowball(
  snowball,
  date,
  force.install = FALSE,
  force.source = FALSE,
  quiet.install = TRUE
)
}
\arguments{
\item{snowball}{dataframe generated by get.snowball() which has the list of packages to be
loaded, whether they need to be installed, where the files will be obtained from, & their installation time.}

\item{date}{character string  (yyyy-mm-dd), or date value, with the date which determines the
version of the package, and all dependencies, to be loaded (and installed if needed).}

\item{force.install}{logical (defaults to \code{FALSE}). When set to \code{TRUE}, will delete
existing package files in groundhog folder, and install anew.}

\item{force.source}{logical (defaults to \code{FALSE}). When set to \code{TRUE}, will not attempt
installing binary from CRAN or MRAN and instead download source file and install it.}

\item{quiet.install}{logical, defaults to \code{TRUE}. When set to \code{FALSE}, displays
output generated by \code{install.packages()} when installing from source}
}
\description{
Install given \code{snowball} from CRAN binary, MRAN binary, CRAN source or MRAN source as needed
}
\seealso{
\code{\link[=get.snowball]{get.snowball()}} to determine in which order packages are installed
}
