% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.snowball.R
\name{install.snowball}
\alias{install.snowball}
\title{Install snowball}
\usage{
install.snowball(
  pkg,
  date,
  include.suggests,
  force.install = FALSE,
  force.source = FALSE,
  quiet.install = TRUE,
  current.deps = c("Rcpp", "RcppArmadillo", "BH", "RcppEigen", "StanHeaders",
    "RcppParallel", "RcppProgress")
)
}
\arguments{
\item{pkg}{character. Name of the package to install}

\item{date}{date (in the format "\%Y-\%m-\%d"), Required date for the package.}

\item{include.suggests}{Logical (defaults to \code{FALSE}. Should suggested
packages be installed?}

\item{force.install}{Logical (defaults to \code{FALSE}). If \code{TRUE}, even if
package is found for this R-build, it will reinstall it.}

\item{force.source}{Logical (defaults to \code{FALSE}). If \code{TRUE},` will skip CRAN
and MRAN attempts, download tar.gz, and install from it.}

\item{quiet.install}{Logical (defaults to \code{TRUE}).Run \code{\link[=install.packages]{install.packages()}}
with \code{quiet = TRUE}?}

\item{current.deps}{Dependencies that will install current version regardless
of date.}
}
\value{
Invisible \code{NULL}.
}
\description{
Install given \code{pkg} with the version from \code{date}, making sure the order in
which dependencies are installed doesn't generate conflicts.
}
\examples{
\donttest{
groundhogR:::install.snowball("magrittr", "2018-02-12", include.suggests = FALSE)
}

}
\seealso{
\code{\link[=get.snowball]{get.snowball()}} to determine in which order packages are installed
}
