% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.source.R
\name{install.source}
\alias{install.source}
\title{Install package from source}
\usage{
install.source(
  pkg_vrs,
  lib,
  date,
  force.download = FALSE,
  quiet.install = FALSE
)
}
\arguments{
\item{pkg_vrs}{character. The required package and its version separated by
an underscore.}

\item{lib}{character. Path where to install the package.}

\item{date}{date (in the format "\%Y-\%m-\%d"), Required date for the package.}

\item{force.download}{logical (defaults to \code{FALSE}). If the tarball
containing the package source code already exists locally, should it be
downloaded again?}

\item{quiet.install}{Logical (defaults to \code{TRUE}).Run \code{\link[=install.packages]{install.packages()}}
with \code{quiet = TRUE}?}
}
\description{
Install package from source
}
