% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.feedback.R
\name{installation.feedback}
\alias{installation.feedback}
\title{Installation feedback}
\usage{
installation.feedback(k, date, snowball, start.time, plot.console = TRUE)
}
\arguments{
\item{k}{numeric. Current step in the snowball process.}

\item{date}{date (in the format "\%Y-\%m-\%d"), Required date for the package.}

\item{snowball}{the result of \code{\link[=get.snowball]{get.snowball()}}.}

\item{start.time}{the result of \code{\link[=Sys.time]{Sys.time()}} when the installation process
started}

\item{plot.console}{logical (defaults to \code{TRUE}). Should detailed feedback
about the installation process be printed in the plot console.}
}
\description{
Print progress about the package installation process (number of packages
already installed or remaining, time left, etc.)
}
