% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.cran.toc.R
\name{load.cran.toc}
\alias{load.cran.toc}
\title{Load \code{cran.toc}}
\usage{
load.cran.toc(update.toc = FALSE)
}
\arguments{
\item{update.toc}{logical (defaults to \code{FALSE}). Should \code{cran.toc} be updated
from the server if it already exists locally.}
}
\value{
(invisibly) \code{TRUE}/\code{FALSE} depending on the success/failure of this
operation.
}
\description{
Load a \code{data.frame} listing all CRAN packages, with their dependencies and
publication date.
}
\examples{
load.cran.toc()

}
\seealso{
\link{update_cran.toc_if.needed}
}
