% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.library.R
\name{restore.library}
\alias{restore.library}
\title{Restore version of packages in personal library replaced by groundhog.}
\usage{
restore.library(days)
}
\arguments{
\item{days}{an optional numeric argument used to choose among alternative restore points.
When \code{days} is set, groundhog restores the personal library to the  most recent restore point, that
is at least \code{days} days ago. For example, if there are two restore points: one from today, and one from 7 days ago,
setting \code{days=3} would restore to the latter, and setting \code{days=8} would result in an error.
\code{days = -1} will restore back to the oldest restore point available.}
}
\description{
When groundhog installs a package, it installs it into  groundhog's library (folder/directory).
Tto find its location: \code{get.groundhog.folder()}. Groundhog then immediately
moves the installed package (and its dependencies) to the default personal library
(to find its location: \code{.libPaths()[1]}). Because the personal library can only
hold one version of a given package, before moving the new package in, groundhog moves any existing
version of those new packages out of the personal library. More generally. Even after packages are installed,
groundhog moves packages in and out of the personal library whenever \code{groundhog.library()} requests a package
version which is not currently there (e.g., when switching back and forth between groundhog days).
\code{restore.library()} allows instantaneously reversing these changes.
It restores the default personal library, removing packages installed by groundhog,
and restoring packages removed by groundhog. This allows, for instance testing out groundhog, and
reversing any changed after doing so.
Groundhog makes a single restore point for every day that \code{groundhog.library()}
leads to changes in the personal library. By default \code{restore.library} reverses to the most recent restore date.
The \code{days} optional argument allows switching to earlier dates (if any).
View other available restore dates with \code{.available.restore.points}.
}
\details{
For more information about groundhog check out \href{http://groundhogr.com}{groundhogr.com}
}
\examples{
\dontrun{
restore.library()
restore.library(7)
restore.library(-1)
}

}
