% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toc.R
\name{toc}
\alias{toc}
\title{Show table of contents (toc) (package versions and publication dates) for a pkg, sorted chronologically}
\usage{
toc(pkg, dependencies = FALSE)
}
\arguments{
\item{pkg}{(required) package name}

\item{dependencies}{logical (defaults to \code{FALSE}). Should the output contain
package dependencies (\code{Imports}, \code{Depends} and \code{Suggests}) for \code{pkg}.}
}
\value{
a \code{data.frame} where each row corresponds to one version of \code{pkg}.
}
\description{
Show table of contents (toc) (package versions and publication dates) for a pkg, sorted chronologically
}
\examples{
toc("magrittr")
}
